# clients/openapi.yaml
openapi: 3.1.0
info:
  title: mem-dog API
  version: 4.0.0
  description: >
    Private AI system for multi-channel data ingestion, AI enrichment,
    temporal knowledge graph, and multi-signal semantic search.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key (md_*), JWT, or global API key
...
paths:
...
  /api/v1/data/{data_id}:
    get:
      tags: [Data]
      summary: Get data content
      operationId: getData
      parameters:
        - name: data_id
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: query
          schema:
            type: integer
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Data content
    put:
      tags: [Data]
      summary: Update data (creates new version)
      operationId: updateData
      parameters:
        - name: data_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                content:
                  type: string
      responses:
        '200':
          description: Updated
    delete:
      tags: [Data]
      summary: Delete data item
      operationId: deleteData
      parameters:
        - name: data_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
...
  /api/v1/ai/embeddings:
    get:
      tags: [Embeddings]
      summary: List embeddings
      operationId: listEmbeddings
      parameters:
        - name: data_id
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: project_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Embeddings
    post:
      tags: [Embeddings]
      summary: Create embedding
      operationId: createEmbedding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data_id:
                  type: string
                engine_type:
                  type: string
                model:
                  type: string
              required: [data_id]
      responses:
        '201':
          description: Created
  /api/v1/ai/embeddings/bulk-delete:
    post:
      tags: [Embeddings]
      summary: Bulk delete embeddings
      operationId: bulkDeleteEmbeddings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Deleted
  /api/v1/ai/embeddings/data/{data_id}:
    get:
      tags: [Embeddings]
      summary: Get data's embeddings
      operationId: getDataEmbeddings
      parameters:
        - name: data_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Embeddings
    delete:
      tags: [Embeddings]
      summary: Delete data's embeddings
      operationId: deleteDataEmbeddings
      parameters:
        - name: data_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /api/v1/ai/embeddings/{embedding_id}:
    get:
      tags: [Embeddings]
      summary: Get embedding
      operationId: getEmbedding
      parameters:
        - name: embedding_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Embedding
    delete:
      tags: [Embeddings]
      summary: Delete embedding
      operationId: deleteEmbedding
      parameters:
        - name: embedding_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted

  # ---------- Viewpoints ----------
...